+++ decidim-admin/app/controllers/decidim/admin/components_controller.rb
@@ -23,11 +23,11 @@ module Decidim
           participatory_space: current_participatory_space
         )
 
-        @form = form(ComponentForm).from_model(@component)
+        @form = form(@component.form_class).from_model(@component)
       end
 
       def create
-        @form = form(ComponentForm).from_params(component_params)
+        @form = form(manifest.component_form_class).from_params(component_params)
         enforce_permission_to :create, :component
 
         CreateComponent.call(@form) do
@@ -47,12 +47,12 @@ module Decidim
         @component = query_scope.find(params[:id])
         enforce_permission_to :update, :component, component: @component
 
-        @form = form(ComponentForm).from_model(@component)
+        @form = form(@component.form_class).from_model(@component)
       end
 
       def update
         @component = query_scope.find(params[:id])
-        @form = form(ComponentForm).from_params(component_params)
+        @form = form(@component.form_class).from_params(component_params)
         enforce_permission_to :update, :component, component: @component
 
         UpdateComponent.call(@form, @component) do
@@ -113,7 +113,7 @@ module Decidim
 
       private
 
-      # Processes the component params so Decidim::Admin::ComponentForm
+      # Processes the component params so the form object defined in the manifest (component_form_class_name)
       # can assign and validate the attributes when using #from_params.
       def component_params
         new_settings = proc { |name, data| Component.build_settings(manifest, name, data, current_organization) }
